+++ probe/procspy/benchmark_internal_test.go
