+++ probe/procspy/fixture.go
