+++ probe/procspy/lsof_internal_test.go
