+++ probe/procspy/netstat_internal_test.go
