+++ probe/procspy/spy.go
